requirements.txt
python-pptx>=0.6.21
pydantic>=2.0.0
python-dateutil>=2.8.0
Pillow>=9.5.0
matplotlib>=3.8.0
jsonschema>=4.18.0
